Wrote WTC. Modified Air Trans classes.

+++ TransportationLab/src/Transportation.java
@@ -25,7 +25,7 @@ public class Transportation {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
